[TA-68] plus propre
