--- Together/app.json
@@ -1,5 +1,0 @@
-{
-  "plugins": [
-    "expo-router"
-  ]
-}
