--- Together/constants/colors.ts
@@ -1,3 +1,0 @@
-export const COLORS = {
-  light_orange: "#FF9A66",
-};
