+++ app/index.tsx
@@ -1,6 +1,9 @@
 import React from 'react';
-import Connexion from '@/pages/Connexion';
+import CustomButton from '@/components/ImageButton'
 
 export default function Index() {
-    return <Connexion />;
+    return <CustomButton
+            image={require("../assets/images/reward.png")}
+            onPress={() => console.log("clic !")}
+            />;
 }
